+++ GalnetMonitor/News.cs
@@ -11,13 +11,17 @@ namespace GalnetMonitor
     /// </summary>
     public class News
     {
+        public string uuid { get; private set; }
         public DateTime published { get; private set; }
+        public string category { get; private set; }
         public string title { get; private set; }
         public string content { get; private set; }
 
-        public News(DateTime published, string title, string content)
+        public News(string uuid, DateTime published, string category, string title, string content)
         {
+            this.uuid = uuid;
             this.published = published;
+            this.category = category;
             this.title = title;
             this.content = content;
         }